+++ docs/guide/changelog.md
@@ -4,6 +4,10 @@
 
 对于新功能、新组件、`bug` 修复以及文档更新，您可以向 `main` 分支创建拉取请求或通过右下角邮箱地址联系我
 
+## <VersionDateTag date="2025-03-05">2.2.13</VersionDateTag>
+
+- 优化并更新 [分页 Pagination](https://themusecatcher.github.io/vue-amazing-ui/guide/components/pagination.html) 组件，新增主题色样式变量，用于支持自定义主题色；新增 `jumperProps` `changerProps` 属性，支持自定义设置 `jumper` 和 `changer` 组件的属性
+
 ## <VersionDateTag date="2025-03-04">2.2.12</VersionDateTag>
 
 - 优化并更新 [按钮 Button](/guide/components/button.html) 组件，新增颜色样式变量，用于支持自定义主题色或样式